CXP-999: add authentication consent wizard
