+++ src/Akeneo/Connectivity/Connection/front/src/connect/hooks/use-confirm-authorization.ts
@@ -10,7 +10,9 @@ interface ConfirmReturn {
 type hookType = (clientId: string) => () => Promise<ConfirmReturn>;
 
 export const useConfirmAuthorization: hookType = clientId => {
-    const url = useRoute('akeneo_connectivity_connection_apps_rest_confirm_authorization', {clientId: clientId, hasUserAuthenticationConsent: '1'});
+    const url = useRoute('akeneo_connectivity_connection_apps_rest_confirm_authorization', {
+        clientId: clientId,
+    });
 
     return useCallback(async () => {
         const response = await fetch(url, {
