+++ src/Akeneo/Connectivity/Connection/front/src/model/Apps/wizard-data.ts
@@ -4,4 +4,5 @@ export interface AppWizardData {
     appName: string;
     appLogo: string;
     scopeMessages: ScopeMessage[];
+    authenticationScopes: Array<'openid' | 'email' | 'profile'>;
 }
