+++ src/Akeneo/Connectivity/Connection/front/tests/src/connect/components/AppWizardWithSteps/Permissions.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
-import {Permissions} from '@src/connect/components/AppWizardWithSteps/Permissions';
+import {Permissions} from '@src/connect/components/AppWizard/steps/Permissions';
 import {renderWithProviders} from '../../../../test-utils';
 import {PermissionsForm} from '@src/connect/components/PermissionsForm';
